src/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3==1.28.57
botocore==1.31.57
click==8.1.7
fastapi==0.103.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.23.2
requests
mangum==0.17.0
